replace usize pinning with CoreId pinning

+++ multicore/src/pin.rs
@@ -3,9 +3,9 @@ use nix::unistd::Pid;
 use nix::Error;
 
 // Pin current thread to specific CPU core 
-pub fn pin_thread_to_core(core: usize) -> Result<(), Error> {
+pub fn pin_thread_to_core(core: u32) -> Result<(), Error> {
     let mut cpu_set = CpuSet::new();
-    cpu_set.set(core)?;
+    cpu_set.set(core as usize)?;
     sched_setaffinity(Pid::from_raw(0), &cpu_set)?;
     Ok(())
 }
